Import ci scoreboard tool.
This is a simple set of scripts that can listen to gerrit, track 3rd
party ci systems reviews, and display a web dashboard showing their
results.

A demo system can be found here:

http://ec2-54-67-102-119.us-west-1.compute.amazonaws.com:5000/?project
=openstack%2Fcinder&user=&timeframe=24

It is not intended to do much more than just show success/fail, and can
work nicely as a quick way to compare your systems results with others
to catch any issues you might be having.

Change-Id: I682c5426fe834a63d3e4f27ebde7d40ee7f9749b