Add some Fibre Channel helper scripts
Credit for these should go to Ramy Asselin. I borrowed the original
versions of these and modified them a bit to work in my system. The
largest change was in the invoke-cf-passthrough function which now is a
standalone script that looks for environment variables for some
parameters. In addition instead of using the gateway x.x.x.1 address to
connect to a hypervisor it will check nova for the host and then connect
with its hostname. This allows it to work for systems using neutron that
cannot ssh through the gateway address. In addition it will try to add
all fc pci devices and not stop after one success. This means if you
are set up for multipathing on the host with multiple hba’s each one
will get passed through to the test node.

The general flow of how to use these is to run the setup script on a
nodepool provider first. Then as part of the jenkins job (or nodepool
ready script) run the passthrough script. At some point in the
process you need to install the fc driver, this should work fine either
as part of the image build process or any point before doing the
passthrough.

Co-Authored-by: Ramy Asselin ramy.asselin@hp.com

Change-Id: Ieba77e99f9d2949f92060483deb58975324e770c