Add in some configurable retries and sleep commands to FC passthrough
Sometimes I get failures passing through the FC devices, it seems that
retrying and adding some slight delay between detaching and attaching
helps with this. If the delay doesn’t fix it, trying a few times will.

They default to 0 second sleeps and up to 3 attempts.

I also added a bit more verbose echo outputs to make it easier to
debug and see what is going on when reading console.log.

Change-Id: I2de5715989bb8a07e54b2dfcbbeae628c7f422c2