Switch the Scoreboard stats into charts
This makes it more easy to 'at a glance' see what a system is up to. Its using
the same counts of success, fail, and unknown as before but now using
EZBC-js to display them as a vertical bar chart. In addition it shows how
many results are 'missing', whether that be that the system missed the
patcheset or just hasn't voted yet.

Change-Id: Ia9d57a74754d64c9faf3826e78ddfb9db574deb7