Add lastcomment
Add lastcomment dashboad to the collection of third party CI tools so it
can be worked on via the standard gerrit workflow.

lastcomment is an alternative monitoring scoreboard that uses gerrit's
REST API as the data source. It was designed to address the developer's
question, is CI system X running/reliable or not?

original source: https://github.com/jogo/lastcomment

Change-Id: I6bb0f321e1885999c9e20f74417dad8768c24c52