Add proof of concept puppet_ci_scoreboard module
This is a first draft of the puppet module to be used to deploy
ci-dashboard.openstack.org as described in [1].

I did not put it into a separate repository on github due to some
restrictions. Adding it here should help to collaborate using
standard review process. If we are happy with the module, we would
move it to openstack/puppet-ci_scoreboard after spec [1] is approved.

[1] https://review.openstack.org/#/c/194437/

Change-Id: I1f0311d081ef970a767470b9e5d2b0a61ec2c8e3