Support for neutron ip’s in clean_floating_ip.sh
Using a liberty based openstack provider with nodepool it seems that
the ‘nova floating-ip-delete <uuid>’ command cannot delete the neutron
floating ip’s.

This will now give the opportunity for a user to specify via environment
variable OS_NETWORK_PROVIDER whether they are using nova or neutron. If
let unset the behavior is unchanged. If neutron is specified it will
make the appropriate neutron floatingip-delete call.

Change-Id: I65f573c9f0a4f11caaa7b673dc1c0f678b64ff9c