[WIP] Initial commit of MoltenIron
MoltenIron is a tool to manage a pool of baremetal nodes that
are to be used as test targets in a baremetal CI environment,
instead of VM guests.

MoltenIron allows you to add, allocate, and release nodes from
it's pool using the following methods:

add - Add a node to the pool

allocate - checkout a baremetal node from the pool, returning
the required info in json format to the requester. It then
marks the node as in use so that no other VM will check it out.

release - return the baremetal node to the pool, allowing another
VM to eventually allocate it.

Change-Id: I8d276d677d9b09bc34032f46c825320d5d83e756