Adds the new cireporter script used by Cinder team
This patch creates the cireporter script which is used by the Cinder
team.   This script is used to help determine which Vendor CI's are working
and reporting over a release cycle.

This patch refactors some of the code in the lastcomment.py into a common
file that contains the Job and Comment classes used by lastcomment and
cireporter.  Also updated the ci.yaml.

Change-Id: I0bde0f539d0e3752963594ef60c3ed460c918116