Fix up help message in cireporter.py
Minor fix to the help message to indicate count
(it was just a copy paste of Name).

Also added an example to the -p argument to show
that the openstack domain is needed when specifying
a project.

Lastly, either capitilize the first work in the help
string or don't, but be consistent.  More were LC so
I made the rest match.

Change-Id: I89e88e51459583c2564b033f7f45223eae77f2c1